fix: sticky adlibs will follow toBeQueued property of pieces
(cherry-picked from a65d875c on olzzon/Core)
